+++ RTClib.h
@@ -40,7 +40,7 @@ protected:
 class TimeSpan {
 public:
     TimeSpan (int32_t seconds = 0);
-    TimeSpan (int16_t days, int8_t hours, int8_t minutes, int8_t seconds);
+    TimeSpan (int32_t days, int32_t hours, int32_t minutes, int32_t seconds);
     TimeSpan (const TimeSpan& copy);
     int16_t days() const         { return _seconds / 86400L; }
     int8_t  hours() const        { return _seconds / 3600 % 24; }
